+++ AzeriteUI/Components/UnitFrames/Auras/Filters.lua
@@ -26,14 +26,15 @@
 local Addon, ns = ...
 ns.AuraFilters = ns.AuraFilters or {}
 
+-- https://wowpedia.fandom.com/wiki/API_C_UnitAuras.GetAuraDataByAuraInstanceID
 ns.AuraFilters.PlayerAuraFilter = function(button, unit, data)
 
 	--button.unitIsCaster = unit and caster and UnitIsUnit(unit, caster)
 	button.spell = data.name
-	button.timeLeft = data.expiration and (data.expiration - GetTime())
-	button.expiration = data.expiration
+	button.timeLeft = data.expirationTime - GetTime()
+	button.expiration = data.expirationTime
 	button.duration = data.duration
-	button.noDuration = (not data.duration or data.duration == 0)
+	button.noDuration = not data.duration or data.duration == 0
 	button.isPlayer = data.isPlayerAura
 
 	if (data.isBossDebuff) then
@@ -41,9 +42,9 @@ ns.AuraFilters.PlayerAuraFilter = function(button, unit, data)
 	end
 
 	if (UnitAffectingCombat("player")) then
-		return (not button.noDuration and data.duration < 301) or (button.timeLeft and button.timeLeft > 0 and button.timeLeft < 31) or (data.count and data.count > 1)
+		return (not button.noDuration and data.duration < 301) or (button.timeLeft and button.timeLeft > 0 and button.timeLeft < 31) or (data.applications > 1)
 	else
-		return (not button.noDuration) or (button.timeLeft and button.timeLeft > 0 and button.timeLeft < 31) or (data.count and data.count > 1)
+		return (not button.noDuration) or (button.timeLeft and button.timeLeft > 0 and button.timeLeft < 31) or (data.applications > 1)
 	end
 
 end
@@ -51,10 +52,10 @@ end
 ns.AuraFilters.TargetAuraFilter = function(button, unit, data)
 
 	button.spell = data.name
-	button.timeLeft = data.expiration and (data.expiration - GetTime())
-	button.expiration = data.expiration
+	button.timeLeft = data.expirationTime - GetTime()
+	button.expiration = data.expirationTime
 	button.duration = data.duration
-	button.noDuration = (not data.duration or data.duration == 0)
+	button.noDuration = not data.duration or data.duration == 0
 	button.isPlayer = data.isPlayerAura
 
 	if (data.isBossDebuff) then
@@ -62,48 +63,50 @@ ns.AuraFilters.TargetAuraFilter = function(button, unit, data)
 	end
 
 	if (UnitAffectingCombat("player")) then
-		return (not button.noDuration and data.duration < 301) or (data.count and data.count > 1)
+		return (not button.noDuration and data.duration < 301) or (data.applications > 1)
 	else
-		return (not button.noDuration) or (data.count and data.count > 1)
+		return (not button.noDuration) or (data.applications > 1)
 	end
 end
 
 ns.AuraFilters.PartyAuraFilter = function(button, unit, data)
 
 	button.spell = data.name
-	button.timeLeft = data.expiration and (data.expiration - GetTime())
-	button.expiration = data.expiration
+	button.timeLeft = data.expirationTime - GetTime()
+	button.expiration = data.expirationTime
 	button.duration = data.duration
-	button.noDuration = (not data.duration or data.duration == 0)
+	button.noDuration = not data.duration or data.duration == 0
 	button.isPlayer = data.isPlayerAura
 
 	if (UnitAffectingCombat("player")) then
-		return (not button.noDuration and data.duration < 301) or (data.count and count > 1)
+		return (not button.noDuration and data.duration < 301) or (data.applications > 1)
 	else
-		return (not button.noDuration) or (data.count and count > 1)
+		return (not button.noDuration) or (data.applications > 1)
 	end
 end
 
 ns.AuraFilters.NameplateAuraFilter = function(button, unit, data)
 
 	button.spell = data.name
-	button.timeLeft = data.expiration and (data.expiration - GetTime())
-	button.expiration = data.expiration
+	button.timeLeft = data.expirationTime - GetTime()
+	button.expiration = data.expirationTime
 	button.duration = data.duration
-	button.noDuration = (not data.duration or data.duration == 0)
+	button.noDuration = not data.duration or data.duration == 0
 	button.isPlayer = data.isPlayerAura
 
 	if (data.isBossDebuff) then
 		return true
 	elseif (data.isStealable) then
 		return true
-	elseif (data.nameplateShowAll) then
-		return true
-	elseif (data.nameplateShowSelf and button.isPlayer) then
+	elseif (data.isNameplateOnly or data.nameplateShowAll or (data.nameplateShowPersonal and button.isPlayer)) then
 		if (button.isHarmful) then
-			return (not button.noDuration and data.duration < 61) or (data.count and data.count > 1)
+			return (not button.noDuration and data.duration < 61) or (data.applications > 1)
 		else
-			return (not button.noDuration and data.duration < 31) or (data.count and data.count > 1)
+			return (not button.noDuration and data.duration < 31) or (data.applications > 1)
+		end
+	else
+		if (not button.isHarmful and button.isPlayer and data.canApplyAura) then
+			return (not button.noDuration and data.duration < 31) or (data.applications > 1)
 		end
 	end
 end